requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
PyJWT>=2.8.0
bcrypt>=4.1.0
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-mock>=3.0.0
hypothesis>=6.0.0
